[ui]: 학기 성적 테이블, 상세성적테이블 컴포넌트 생성 및 분리 (#14)
